Add ability to log detailed Fedora debug messages via a virtio console
Some early boot failures are visible only on the console and details
are only visible if certain additional kernel command line options are
passed.

This patch adds the required kernel command lines, redirects console to
a serial port and then monitors both the regular serial and virtio ports
during install and logs full line output.

At the time of this commit there is no way to combine these log streams
into a single location.

This also changes the serial monitoring socket from a TCP connection to a
domain socket file.  Without this I ran into issues with the listening socket
failing to recreate when doing ICICLE generation due to TCP stack timeout
issues.

(This is a squashed and cherry-picked version of a series of forked commits
rebased onto the 0.15 release.)